refactor(ProductService): refatorando service para passar nos testes unitários

+++ src/main/java/com/produto/api/service/ProductService.java
@@ -13,6 +13,7 @@ import jakarta.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Service
@@ -23,6 +24,13 @@ public class ProductService {
     ProductMapper mapper;
 
     public void addProduct(AddProductDTO product) {
+        if (product.quantity() == null || product.quantity() < 0) {
+            throw new IllegalArgumentException();
+        } else if (product.name() == null || product.name().isEmpty()) {
+            throw new IllegalArgumentException();
+        } else if(product.price() == null || product.price().compareTo(BigDecimal.ZERO) < 0) {
+            throw new IllegalArgumentException();
+        }
         repository.save(mapper.toEntityAdd(product));
     }
 
@@ -32,23 +40,32 @@ public class ProductService {
     }
 
     public ResponseProductDTO findById(Long id){
+        if(id == null) throw new IllegalArgumentException();
         if (repository.findById(id).isEmpty()) throw new ProductNotFoundException("Product with id " + id + " not found");
         return mapper.toDTO(repository.findById(id).get());
     }
 
     public void deleteProduct(Long id){
+        if (id == null) throw new IllegalArgumentException();
         if (repository.findById(id).isEmpty()) throw new ProductNotFoundException("Product with id " + id + " not found");
         repository.deleteById(id);
     }
 
     public void updateProduct(Long id, UpdateProductDTO updatedProduct) {
+        if (id == null) throw new IllegalArgumentException();
+        if (updatedProduct.name() == null || updatedProduct.name().isEmpty()) throw new IllegalArgumentException();
+        if (updatedProduct.quantity() == null || updatedProduct.quantity() < 0) throw new IllegalArgumentException();
+        if (updatedProduct.price() == null || updatedProduct.price().compareTo(BigDecimal.ZERO) < 0) throw new IllegalArgumentException();
         if (repository.findById(id).isEmpty()) throw new ProductNotFoundException("Product with id " + id + " not found");
+
         Product produto = repository.findById(id).get();
         mapper.toEntityUpdate(updatedProduct, produto);
         repository.save(produto);
     }
 
     public void withdrawProduct(Long id, WithdrawOrPutProductDTO withdrawProduct) {
+        if (id == null) throw new IllegalArgumentException();
+        if (withdrawProduct.quantity() == null || withdrawProduct.quantity() < 0) throw new IllegalArgumentException();
         if (repository.findById(id).isEmpty()) throw new ProductNotFoundException("Product with id " + id + " not found");
         Product produto = repository.findById(id).get();
         if (produto.getQuantity() == 0 || produto.getQuantity() < withdrawProduct.quantity()) throw new NotEnoghProductException("Not enough products in stock");
@@ -56,7 +73,9 @@ public class ProductService {
         repository.save(produto);
     }
 
-    public void putProduct(Long id, @Valid UpdateProductDTO putProduct) {
+    public void putProduct(Long id, @Valid WithdrawOrPutProductDTO putProduct) {
+        if (id == null) throw new IllegalArgumentException();
+        if (putProduct.quantity() == null || putProduct.quantity() < 0) throw new IllegalArgumentException();
         if (repository.findById(id).isEmpty()) throw new ProductNotFoundException("Product with id " + id + " not found");
         Product produto = repository.findById(id).get();
         produto.setQuantity(produto.getQuantity()+putProduct.quantity());
